Fix render profile image

+++ Admin_view/categoryList.php
@@ -8,23 +8,23 @@ include './include/head.php';
 <body>
   <?php require_once "../database/config.php";
 
-    if (isset($_GET['err_exist'])) {
-        echo '<script type="text/javascript">alert("' . $_GET['err_exist'] . '");</script>';
-    }
-    if (isset($_GET['remove_success'])) {
-        echo '<script type="text/javascript">alert("' . $_GET['remove_success'] . '");</script>';
-    }
-    ?>
+  if (isset($_GET['err_exist'])) {
+    echo '<script type="text/javascript">alert("' . $_GET['err_exist'] . '");</script>';
+  }
+  if (isset($_GET['remove_success'])) {
+    echo '<script type="text/javascript">alert("' . $_GET['remove_success'] . '");</script>';
+  }
+  ?>
   <!-- HEADER -->
   <?php
-    include './include/header.php';
-    ?>
+  include './include/header.php';
+  ?>
   <div class="container-fluid">
     <div class="row">
       <!-- NAVIGATION -->
       <?php
-            include './include/nav.php';
-            ?>
+      include './include/nav.php';
+      ?>
       <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4">
 
         <div class="container-fluid">
@@ -32,10 +32,8 @@ include './include/head.php';
             <div class="d-flex flex-row justify-content-between mt-4">
               <div class="d-flex flex-row">
                 <input type="hidden" id="page" name="page" class="form-control">
-                <input type="text" id="keyword" name="keyword" class="form-control" placeholder="Search" value=""
-                  style="margin-right: 5px;">
-                <button type="submit" id="btnSearch" name="btnSearch" value="Search"
-                  class="btn btn-primary">Seach</button>
+                <input type="text" id="keyword" name="keyword" class="form-control" placeholder="Search" value="" style="margin-right: 5px;">
+                <button type="submit" id="btnSearch" name="btnSearch" value="Search" class="btn btn-primary">Seach</button>
               </div>
               <div>
                 <a class="btn btn-outline-primary mb-1" href="./addcategory.php" role="button">Add
@@ -51,24 +49,19 @@ include './include/head.php';
               </thead>
               <tbody>
                 <?php
-                                $sql = "SELECT * FROM `categories`";
-                                $categories = $mysqli->query($sql);
-                                mysqli_close($mysqli);
-                                ?>
+                $sql = "SELECT * FROM `categories`";
+                $categories = $mysqli->query($sql);
+                mysqli_close($mysqli);
+                ?>
                 <?php foreach ($categories as $category) { ?>
-                <tr>
-                  <th scope="row"><?php echo $category['category_id'] ?></th>
-                  <td><?php echo $category['category_name'] ?></td>
-                  <td>
-                    <a class="btn btn-primary"
-                      href="./category/update_category.php?id=<?php echo $category['category_id'] ?>"
-                      role="button">Edit</a>
-                    <a class="btn btn-danger"
-                      href="./category/process_deletecategory.php?id=<?php echo $category['category_id'] ?>"
-                      onclick="return confirm('Bạn có muốn xóa <?php echo $category['category_name'] ?>')"
-                      role="button">Delete</a>
-                  </td>
-                </tr>
+                  <tr>
+                    <th scope="row"><?php echo $category['category_id'] ?></th>
+                    <td><?php echo $category['category_name'] ?></td>
+                    <td>
+                      <a class="btn btn-primary" href="./category/update_category.php?id=<?php echo $category['category_id'] ?>" role="button">Edit</a>
+                      <a class="btn btn-danger" href="./category/process_deletecategory.php?id=<?php echo $category['category_id'] ?>" onclick="return confirm('Bạn có muốn xóa <?php echo $category['category_name'] ?>')" role="button">Delete</a>
+                    </td>
+                  </tr>
                 <?php } ?>
               </tbody>
             </table>
@@ -92,29 +85,27 @@ include './include/head.php';
 
   <script src="https://getbootstrap.com/docs/5.1/dist/js/bootstrap.bundle.min.js"></script>
   <script src="//netdna.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/feather-icons@4.28.0/dist/feather.min.js"
-    integrity="sha384-uO3SXW5IuS1ZpFPKugNNWqTZRRglnUJK6UAZ/gxOX80nxEkN9NcGZTftn6RzhGWE" crossorigin="anonymous">
+  <script src="https://cdn.jsdelivr.net/npm/feather-icons@4.28.0/dist/feather.min.js" integrity="sha384-uO3SXW5IuS1ZpFPKugNNWqTZRRglnUJK6UAZ/gxOX80nxEkN9NcGZTftn6RzhGWE" crossorigin="anonymous">
   </script>
-  <script src="https://cdn.jsdelivr.net/npm/chart.js@2.9.4/dist/Chart.min.js"
-    integrity="sha384-zNy6FEbO50N+Cg5wap8IKA4M/ZnLJgzc6w2NqACZaK0u0FXfOWRRJOnQtpZun8ha" crossorigin="anonymous">
+  <script src="https://cdn.jsdelivr.net/npm/chart.js@2.9.4/dist/Chart.min.js" integrity="sha384-zNy6FEbO50N+Cg5wap8IKA4M/ZnLJgzc6w2NqACZaK0u0FXfOWRRJOnQtpZun8ha" crossorigin="anonymous">
   </script>
   <script src="${base}/jsdashboard.js"></script>
 
 
   <script type="text/javascript">
-  // $( document ).ready(function() {
-  // 	$("#categoryId").val(${searchModel.categoryId});
-  // 	$("#paging").pagination({
-  // 		currentPage: ${CategoryWithPaging.currentPage},
-  //         items: ${CategoryWithPaging.totalItems},
-  //         itemsOnPage: 4,
-  //         cssStyle: 'light-theme',
-  //         onPageClick: function(pageNumber, event) {
-  //         	$('#page').val(pageNumber);
-  //         	$('#btnSearch').trigger('click');
-  // 		},
-  //     });
-  // });
+    // $( document ).ready(function() {
+    // 	$("#categoryId").val(${searchModel.categoryId});
+    // 	$("#paging").pagination({
+    // 		currentPage: ${CategoryWithPaging.currentPage},
+    //         items: ${CategoryWithPaging.totalItems},
+    //         itemsOnPage: 4,
+    //         cssStyle: 'light-theme',
+    //         onPageClick: function(pageNumber, event) {
+    //         	$('#page').val(pageNumber);
+    //         	$('#btnSearch').trigger('click');
+    // 		},
+    //     });
+    // });
   </script>
   <%--  <script type="text/javascript" src="${base}/js/jquery.js"></script> --%>
 </body>
