+++ Admin_view/user/update_user.php
@@ -1,3 +1,45 @@
+<?php
+require_once "../../database/config.php";
+
+if (isset($_POST['submit'])) {
+  $user_id = $_POST['user_id'];
+  $fullname = $_POST['fullname'];
+  $email = $_POST['email'];
+  $phone = $_POST['phone'];
+  $role = $_POST['role'];
+  $name = $_POST['name'];
+  $avatar = $_FILES['user_avatar']['name'];
+  $avatar_tmp = $_FILES['user_avatar']['tmp_name'];
+  $path = '../upload/user/';
+
+  $sql_check = "SELECT count(*) FROM users where email='$email' and user_id <> '$user_id'";
+  $result = $mysqli->query($sql_check);
+  $number_rows = mysqli_fetch_array($result)['count(*)'];
+
+  if ($number_rows == 1) {
+    header("location: ./update_user.php?id=$user_id&err_exist=Tài khoản đã tồn tại rùi");
+    exit;
+  }
+
+  $sql = "UPDATE users
+          SET 
+          name='$name',
+          phone='$phone',
+          username='$fullname',
+          email='$email',
+          role='$role',
+          avatar='$avatar'
+          WHERE user_id='$user_id'";
+
+  $mysqli->query($sql);
+  move_uploaded_file($avatar_tmp, $path . $avatar);
+
+  $_SESSION['email'] = $email;
+
+  header('location: ../userlist.php?success=true');
+}
+?>
+
 <html lang="en">
 
 <head>
@@ -19,31 +61,31 @@
   <link href="https://getbootstrap.com/docs/5.1/dist/css/bootstrap.min.css" rel="stylesheet">
 
   <style>
-  .bd-placeholder-img {
-    font-size: 1.125rem;
-    text-anchor: middle;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    user-select: none;
-  }
-
-  @media (min-width: 768px) {
-    .bd-placeholder-img-lg {
-      font-size: 3.5rem;
+    .bd-placeholder-img {
+      font-size: 1.125rem;
+      text-anchor: middle;
+      -webkit-user-select: none;
+      -moz-user-select: none;
+      user-select: none;
     }
-  }
 
-  .col-md-4 {
-    width: 93.333333% !important;
-  }
+    @media (min-width: 768px) {
+      .bd-placeholder-img-lg {
+        font-size: 3.5rem;
+      }
+    }
 
-  .cus-info {
-    width: 100%;
-  }
+    .col-md-4 {
+      width: 93.333333% !important;
+    }
 
-  .form-group {
-    margin-bottom: 12px;
-  }
+    .cus-info {
+      width: 100%;
+    }
+
+    .form-group {
+      margin-bottom: 12px;
+    }
   </style>
 
   <!-- Custom styles for this template -->
@@ -54,14 +96,10 @@
 <body>
   <?php
   include '../include/header.php';
-  require_once "../../database/config.php";
   ?>
   <div>
-    <?php
-    include '../include/nav.php';
-    ?>
+    <?php include '../include/nav.php'; ?>
   </div>
-
   <?php
   if (isset($_GET['err_exist'])) {
     echo '<script type="text/javascript">alert("' . $_GET['err_exist'] . '");</script>';
@@ -78,23 +116,21 @@
         ?>
 
         <div class="container">
-          <form class="form-inline" method="POST" action="./process_user.php">
+          <form class="form-inline" method="POST" action="" enctype="multipart/form-data">
             <div class="main-body">
               <div class="row gutters-sm">
                 <div class="col-md-4 mb-3">
                   <div class="card">
                     <div class="card-body">
                       <div class="d-flex flex-column align-items-center text-center">
-                        <img src="https://bootdey.com/img/Content/avatar/avatar7.png" alt="Admin" class="rounded-circle"
-                          width="150">
+                        <img src="/PhuongNamSport/Admin_view/upload/user/<?php echo $user['avatar'] ?>" alt="Admin" class="rounded-circle" width="150" height="150">
                         <div class="mt-3">
                           <h4><?php echo $user['username'] ?></h4>
                         </div>
                       </div>
                       <div class="form-group">
                         <label for="inputAddress">Full Name</label>
-                        <input type="text" class="form-control" id="inputAddress" placeholder="Enter your fullname"
-                          value="<?php echo $user['username'] ?>" name="fullname">
+                        <input type="text" class="form-control" id="inputAddress" placeholder="Enter your fullname" value="<?php echo $user['username'] ?>" name="fullname">
                       </div>
 
                       <!-- Hidden field -->
@@ -102,18 +138,24 @@
 
                       <div class="form-group">
                         <label for="inputAddress2">Email</label>
-                        <input type="text" class="form-control" id="inputAddress2" placeholder="Enter your email"
-                          value="<?php echo $user['email'] ?>" name="email">
+                        <input type="text" class="form-control" id="inputAddress2" placeholder="Enter your email" value="<?php echo $user['email'] ?>" name="email">
                       </div>
                       <div class="form-group">
                         <label for="inputAddress">Phone</label>
-                        <input type="text" class="form-control" id="inputAddress" placeholder="Enter your phone"
-                          value="<?php echo $user['phone'] ?>" name="phone">
+                        <input type="text" class="form-control" id="inputAddress" placeholder="Enter your phone" value="<?php echo $user['phone'] ?>" name="phone">
                       </div>
                       <div class="form-group">
                         <label for="inputAddress2">Name</label>
-                        <input type="text" class="form-control" id="inputAddress2" placeholder="Enter your name"
-                          value="<?php echo $user['name'] ?>" name="name">
+                        <input type="text" class="form-control" id="inputAddress2" placeholder="Enter your name" value="<?php echo $user['name'] ?>" name="name">
+                      </div>
+
+                      <!-- phần thêm avatar   -->
+                      <!-- File Button -->
+                      <div class="form-group">
+                        <label class="col-md-4 control-label" for="avatar">Avatar</label>
+                        <div style="width: 93.333333% !important;" class="col-md-4">
+                          <input id="avatarfile" name="user_avatar" class="input-file" type="file" />
+                        </div>
                       </div>
                       <div class="form-group">
                         <label class="col-md-4 control-label" for="category">Role</label>
@@ -128,10 +170,7 @@
                         </div>
                       </div>
                       <button type="submit" name="submit" class="btn btn-primary">Edit</button>
-                      <a href="process_user.php?id_delete=<?php echo $user['user_id'] ?>"
-                        class="btn btn-danger btn-md active"
-                        onclick="return confirm('Bạn chắc chắn muốn xóa tài khoản <?php echo $user['username'] ?> này?')"
-                        role="button" aria-pressed=" true">Remove</a>
+                      <a href="process_user.php?id_delete=<?php echo $user['user_id'] ?>" class="btn btn-danger btn-md active" onclick="return confirm('Bạn chắc chắn muốn xóa tài khoản <?php echo $user['username'] ?> này?')" role="button" aria-pressed=" true">Remove</a>
                     </div>
                   </div>
                 </div>
@@ -146,11 +185,9 @@
 
   <script src="https://getbootstrap.com/docs/5.1/dist/js/bootstrap.bundle.min.js"></script>
 
-  <script src="https://cdn.jsdelivr.net/npm/feather-icons@4.28.0/dist/feather.min.js"
-    integrity="sha384-uO3SXW5IuS1ZpFPKugNNWqTZRRglnUJK6UAZ/gxOX80nxEkN9NcGZTftn6RzhGWE" crossorigin="anonymous">
+  <script src="https://cdn.jsdelivr.net/npm/feather-icons@4.28.0/dist/feather.min.js" integrity="sha384-uO3SXW5IuS1ZpFPKugNNWqTZRRglnUJK6UAZ/gxOX80nxEkN9NcGZTftn6RzhGWE" crossorigin="anonymous">
   </script>
-  <script src="https://cdn.jsdelivr.net/npm/chart.js@2.9.4/dist/Chart.min.js"
-    integrity="sha384-zNy6FEbO50N+Cg5wap8IKA4M/ZnLJgzc6w2NqACZaK0u0FXfOWRRJOnQtpZun8ha" crossorigin="anonymous">
+  <script src="https://cdn.jsdelivr.net/npm/chart.js@2.9.4/dist/Chart.min.js" integrity="sha384-zNy6FEbO50N+Cg5wap8IKA4M/ZnLJgzc6w2NqACZaK0u0FXfOWRRJOnQtpZun8ha" crossorigin="anonymous">
   </script>
   <script src="../dashboard.js"></script>
   <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
